ファイル構成のリファクタリング
