--- app/javascript/packs/components/uploads/index.vue
@@ -1,14 +1,0 @@
-<template lang="pug">
-  div
-    h1 {{msg}}
-</template>
-
-<script>
-  export default {
-    data () {
-      return {
-        msg: 'This is Upload!!'
-      }
-    }
-  }
-</script>
