--- app/javascript/packs/views/Home/index.vue
@@ -1,12 +1,0 @@
-<template lang="pug">
-  div
-    Audios
-</template>
-
-<script>
-  export default {
-    components: {
-      Audios: require('./home.vue')
-    }
-  }
-</script>
