--- app/javascript/packs/views/Upload/index.vue
@@ -1,12 +1,0 @@
-<template lang="pug">
-  div
-    Upload
-</template>
-
-<script>
-  export default {
-    components: {
-      Upload: require('./upload.vue')
-    }
-  }
-</script>
