--- app/javascript/packs/views/Upload/upload.vue
@@ -1,16 +1,0 @@
-<template lang="pug">
-  div
-    Upload
-    div
-      h2 ダミーテキスト
-</template>
-
-<script>
-  import Upload from '../../components/uploads/index.vue';
-  export default {
-    name: 'upload',
-    components: {
-        Upload
-    }
-}
-</script>
