+++ app/javascript/packs/views/upload.vue
@@ -1,0 +1,15 @@
+<template lang="pug">
+  div
+    UploadForm
+    div
+      h2 ダミーテキストaaaaas
+</template>
+
+<script>
+  import UploadForm from '../components/uploadform.vue';
+  export default {
+    components: {
+        UploadForm
+    }
+}
+</script>
